+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `Stamp<S, Tm>` constant-`dt` source wrapper
 - Added `Filter<f32>` quadrature map for `Nco<T>`
 - Added `Filter<T>` waveform maps for square and sawtooth oscillators
+- Added `Filter<T>` waveform maps for triangle and pulse oscillators
 
 ### Changed